Handle multiple validators with a variable.
Overall, add a `validator` variable which is populated from a label
named `validator` in the prometheus config. This allows easy switching
between validators on the dashboard

Also, several cleanups that are probably optional. They are:
* Changed the title from "miner stats" to "Validator stats". Not sure
  this is actually better.

* Turned on `bars` for penalty panel. I think this makes stacking
  easier to see.

* Added aliases and ledgend overrides for all of the queries to make
  them shorter on the panels.

* Truncated decimal places ("decimals": 0) in many places. I think
  this makes the Y axis easier to understand, especially when switching
  between validators.

* Set a min of 0 on many panels, to keep the y-axis from showing
  negative values in an attempt to vertically center data sometimes.

* Hid the right y-axis (consensus) on the penalty panel. Is really a
  binary value. The bars being present as a sort of overaly is all the
  info we need.

Also, Grafana itself seemed to make some other changes with things
like __inputs, __requires, and hashKey. I have no idea if those are
important, but this seems to import for me.

Finally, I'm running Grafana 7.5.5, but I manually reverted the hunks
that changed 7.5.4 -> 7.5.5 and checked I could re-import it just fine.