Automagically set 'server_name' when setting 'validator_name' variable
When setting the validator_name variable, use a regex on the instance
field to automatically figure out the FQDN and set server_name to that
value. Then, we can use this FQDN in all the queries that use `instance=`
by using a little more regex in those (`instance=~`).

This assume the promethues targets uses the same IP or FQDN for both
the miner_exporter and node_exporter, varying only the port number.

Also, I hid the server_name variable because there shouldn't be more
than one value to choose from, but it might nice to have a panel that
displated the value of $server_name (it might be an IP or FQDN), but
this is already shown in the "validator version" panel, so I didn't
add that in this commit